Add `switch_h1_double.py` and move H1 double switch quirk from `opple_switch.py`
